Kaiserschlacht loca
Also a few more tweaks, namely that taking the border will now add/remove cores on Briey + Nanzig, fixed a typo with the Chinese peace event that would cause it to fire in the middle of the peace chain, tweaked the decline numbers and made the austrians more likely to go to war with Italy should they decline the peace, some ai_chance work, as well as state names if Germany chooses the 1871 border.

Also localization for the kaiserschlacht events. Yay.
A tooltip was also added to warn the German player that declining the peace event will mean having to invade all majors.